requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
click==8.1.3
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iteration-utilities==0.11.0
mypy-extensions==1.0.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.2
random-user-agent==1.0.1
requests==2.28.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.5.0
urllib3==1.26.15
websockets==11.0.3
wsproto==1.2.0
